+++ src/zspec/zcl/definition/clusters-types.ts
@@ -2378,7 +2378,7 @@ export interface TClusters {
             /** ID=0x0002 | type=BITMAP8 | write=true | required=true | default=0 */
             status: number;
             /** ID=0x0010 | type=UINT16 | write=true | required=true | min=1 | max=65534 | default=1 */
-            losedLimit: number;
+            closedLimit: number;
             /** ID=0x0011 | type=ENUM8 | write=true | required=true | max=254 | default=0 */
             mode: number;
         };
@@ -2911,11 +2911,11 @@ export interface TClusters {
             accelerationTimeLift: number;
             /** ID=0x0016 | type=UINT16 | write=true | max=65535 | default=0 */
             decelerationTimeLift: number;
-            /** ID=0x0017 | type=BITMAP8 | required=true | default=4 */
+            /** ID=0x0017 | type=BITMAP8 | write=true | required=true | default=4 */
             windowCoveringMode: number;
-            /** ID=0x0018 | type=OCTET_STR | default=1,0x0000 */
+            /** ID=0x0018 | type=OCTET_STR | write=true | default=1,0x0000 */
             intermediateSetpointsLift: Buffer;
-            /** ID=0x0019 | type=OCTET_STR | default=1,0x0000 */
+            /** ID=0x0019 | type=OCTET_STR | write=true | default=1,0x0000 */
             intermediateSetpointsTilt: Buffer;
             /** ID=0x000a | type=BITMAP8 */
             operationalStatus: number;
@@ -3189,48 +3189,14 @@ export interface TClusters {
             fourNoksHysteresisLow?: number;
             /** ID=0x0400 | type=ENUM8 | manufacturerCode=SINOPE_TECHNOLOGIES(0x119c) | write=true | max=255 */
             SinopeOccupancy?: number;
-            /** ID=0x0401 | type=UINT16 | write=true | max=65535 */
-            elkoLoad: number;
             /** ID=0x0401 | type=UINT16 | manufacturerCode=SINOPE_TECHNOLOGIES(0x119c) | write=true | max=65535 */
             SinopeMainCycleOutput?: number;
             /** ID=0x0402 | type=CHAR_STR | write=true */
             elkoDisplayText: string;
             /** ID=0x0402 | type=ENUM8 | manufacturerCode=SINOPE_TECHNOLOGIES(0x119c) | write=true | max=255 */
             SinopeBacklight?: number;
-            /** ID=0x0403 | type=ENUM8 | write=true | max=255 */
-            elkoSensor: number;
-            /** ID=0x0404 | type=UINT8 | write=true | max=255 */
-            elkoRegulatorTime: number;
             /** ID=0x0404 | type=UINT16 | manufacturerCode=SINOPE_TECHNOLOGIES(0x119c) | write=true | max=65535 */
             SinopeAuxCycleOutput?: number;
-            /** ID=0x0405 | type=BOOLEAN | write=true */
-            elkoRegulatorMode: number;
-            /** ID=0x0406 | type=BOOLEAN | write=true */
-            elkoPowerStatus: number;
-            /** ID=0x0407 | type=OCTET_STR | write=true */
-            elkoDateTime: Buffer;
-            /** ID=0x0408 | type=UINT16 | write=true | max=65535 */
-            elkoMeanPower: number;
-            /** ID=0x0409 | type=INT16 | write=true | min=-32768 | max=32767 */
-            elkoExternalTemp: number;
-            /** ID=0x0411 | type=BOOLEAN | write=true */
-            elkoNightSwitching: number;
-            /** ID=0x0412 | type=BOOLEAN | write=true */
-            elkoFrostGuard: number;
-            /** ID=0x0413 | type=BOOLEAN | write=true */
-            elkoChildLock: number;
-            /** ID=0x0414 | type=UINT8 | write=true | max=255 */
-            elkoMaxFloorTemp: number;
-            /** ID=0x0415 | type=BOOLEAN | write=true */
-            elkoRelayState: number;
-            /** ID=0x0416 | type=OCTET_STR | write=true */
-            elkoVersion: Buffer;
-            /** ID=0x0417 | type=INT8 | write=true | min=-128 | max=127 */
-            elkoCalibration: number;
-            /** ID=0x0418 | type=UINT8 | write=true | max=255 */
-            elkoLastMessageId: number;
-            /** ID=0x0419 | type=UINT8 | write=true | max=255 */
-            elkoLastMessageStatus: number;
             /** ID=0x4000 | type=ENUM8 | manufacturerCode=VIESSMANN_ELEKTRONIK_GMBH(0x1221) | write=true | max=255 */
             viessmannWindowOpenInternal?: number;
             /** ID=0x4000 | type=ENUM8 | manufacturerCode=DANFOSS_A_S(0x1246) | write=true | max=255 */
@@ -6607,9 +6573,9 @@ export interface TClusters {
             model: Buffer;
             /** ID=0x0007 | type=OCTET_STR | minLen=0 | maxLen=16 */
             partNumber: Buffer;
-            /** ID=0x0008 | type=OCTET_STR | minLen=0 | maxLen=16 */
+            /** ID=0x0008 | type=OCTET_STR | minLen=0 | maxLen=6 */
             productRevision: Buffer;
-            /** ID=0x000a | type=OCTET_STR | minLen=0 | maxLen=16 */
+            /** ID=0x000a | type=OCTET_STR | minLen=0 | maxLen=6 */
             softwareRevision: Buffer;
             /** ID=0x000b | type=CHAR_STR | minLen=0 | maxLen=16 */
             utilityName: string;
@@ -7410,27 +7376,6 @@ export interface TClusters {
             };
         };
     };
-    manuSpecificClusterAduroSmart: {
-        attributes: never;
-        commands: {
-            /** ID=0x00 */
-            cmd0: Record<string, never>;
-        };
-        commandResponses: never;
-    };
-    manuSpecificOsram: {
-        attributes: never;
-        commands: {
-            /** ID=0x01 */
-            saveStartupParams: Record<string, never>;
-            /** ID=0x02 */
-            resetStartupParams: Record<string, never>;
-        };
-        commandResponses: {
-            /** ID=0x00 */
-            saveStartupParamsRsp: Record<string, never>;
-        };
-    };
     manuSpecificPhilips: {
         attributes: {
             /** ID=0x0031 | type=BITMAP16 | write=true */
@@ -7563,33 +7508,6 @@ export interface TClusters {
             coldLoadPickupStatus: number;
         };
         commands: never;
-        commandResponses: never;
-    };
-    manuSpecificLegrandDevices: {
-        attributes: never;
-        commands: never;
-        commandResponses: never;
-    };
-    manuSpecificLegrandDevices2: {
-        attributes: never;
-        commands: {
-            /** ID=0x00 */
-            command0: {
-                /** type=BUFFER */
-                data: Buffer;
-            };
-        };
-        commandResponses: never;
-    };
-    manuSpecificLegrandDevices3: {
-        attributes: never;
-        commands: {
-            /** ID=0x00 */
-            command0: {
-                /** type=BUFFER */
-                data: Buffer;
-            };
-        };
         commandResponses: never;
     };
     manuSpecificTuya: {
@@ -7817,14 +7735,6 @@ export interface TClusters {
         };
         commandResponses: never;
     };
-    manuSpecificCentraliteHumidity: {
-        attributes: {
-            /** ID=0x0000 | type=UINT16 | write=true | max=65535 */
-            measuredValue: number;
-        };
-        commands: never;
-        commandResponses: never;
-    };
     manuSpecificSmartThingsArrivalSensor: {
         attributes: never;
         commands: never;
@@ -7832,124 +7742,6 @@ export interface TClusters {
             /** ID=0x01 */
             arrivalSensorNotify: Record<string, never>;
         };
-    };
-    manuSpecificSamsungAccelerometer: {
-        attributes: {
-            /** ID=0x0000 | type=UINT8 | write=true | max=255 */
-            motion_threshold_multiplier: number;
-            /** ID=0x0002 | type=UINT16 | write=true | max=65535 */
-            motion_threshold: number;
-            /** ID=0x0010 | type=BITMAP8 | write=true | max=255 */
-            acceleration: number;
-            /** ID=0x0012 | type=INT16 | write=true | min=-32768 | max=32767 */
-            x_axis: number;
-            /** ID=0x0013 | type=INT16 | write=true | min=-32768 | max=32767 */
-            y_axis: number;
-            /** ID=0x0014 | type=INT16 | write=true | min=-32768 | max=32767 */
-            z_axis: number;
-        };
-        commands: never;
-        commandResponses: never;
-    };
-    tradfriButton: {
-        attributes: never;
-        commands: {
-            /** ID=0x01 */
-            action1: {
-                /** type=UINT8 | max=255 */
-                data: number;
-            };
-            /** ID=0x02 */
-            action2: {
-                /** type=UINT8 | max=255 */
-                data: number;
-            };
-            /** ID=0x03 */
-            action3: {
-                /** type=UINT8 | max=255 */
-                data: number;
-            };
-            /** ID=0x04 */
-            action4: {
-                /** type=UINT8 | max=255 */
-                data: number;
-            };
-            /** ID=0x06 */
-            action6: {
-                /** type=UINT8 | max=255 */
-                data: number;
-            };
-        };
-        commandResponses: never;
-    };
-    schneiderSpecificPilotMode: {
-        attributes: {
-            /** ID=0x0031 | type=ENUM8 | write=true | max=255 */
-            pilotMode: number;
-        };
-        commands: never;
-        commandResponses: never;
-    };
-    manuSpecificSchneiderFanSwitchConfiguration: {
-        attributes: {
-            /** ID=0x0002 | type=UINT8 | write=true | max=255 */
-            ledIndication: number;
-            /** ID=0x0060 | type=UINT8 | write=true | max=255 */
-            ledOrientation: number;
-        };
-        commands: never;
-        commandResponses: never;
-    };
-    sprutVoc: {
-        attributes: {
-            /** ID=0x6600 | type=UINT16 | write=true | max=65535 */
-            voc: number;
-        };
-        commands: never;
-        commandResponses: never;
-    };
-    sprutNoise: {
-        attributes: {
-            /** ID=0x6600 | type=SINGLE_PREC | write=true */
-            noise: number;
-            /** ID=0x6601 | type=BITMAP8 | write=true */
-            noiseDetected: number;
-            /** ID=0x6602 | type=SINGLE_PREC | write=true */
-            noiseDetectLevel: number;
-            /** ID=0x6603 | type=UINT16 | write=true | max=65535 */
-            noiseAfterDetectDelay: number;
-        };
-        commands: never;
-        commandResponses: never;
-    };
-    sprutIrBlaster: {
-        attributes: never;
-        commands: {
-            /** ID=0x00 */
-            playStore: {
-                /** type=UINT8 | max=255 */
-                param: number;
-            };
-            /** ID=0x01 */
-            learnStart: {
-                /** type=UINT8 | max=255 */
-                value: number;
-            };
-            /** ID=0x02 */
-            learnStop: {
-                /** type=UINT8 | max=255 */
-                value: number;
-            };
-            /** ID=0x03 */
-            clearStore: Record<string, never>;
-            /** ID=0x04 */
-            playRam: Record<string, never>;
-            /** ID=0x05 */
-            learnRamStart: Record<string, never>;
-            /** ID=0x06 */
-            learnRamStop: Record<string, never>;
-        };
-        commandResponses: never;
     };
     manuSpecificSiglisZigfred: {
         attributes: {
@@ -8088,147 +7880,6 @@ export interface TClusters {
             };
         };
         commandResponses: never;
-    };
-    manuSpecificAssaDoorLock: {
-        attributes: {
-            /** ID=0x0012 | type=UINT8 | write=true | max=255 */
-            autoLockTime: number;
-            /** ID=0x0013 | type=UINT8 | write=true | max=255 */
-            wrongCodeAttempts: number;
-            /** ID=0x0014 | type=UINT8 | write=true | max=255 */
-            shutdownTime: number;
-            /** ID=0x0015 | type=UINT8 | write=true | max=255 */
-            batteryLevel: number;
-            /** ID=0x0016 | type=UINT8 | write=true | max=255 */
-            insideEscutcheonLED: number;
-            /** ID=0x0017 | type=UINT8 | write=true | max=255 */
-            volume: number;
-            /** ID=0x0018 | type=UINT8 | write=true | max=255 */
-            lockMode: number;
-            /** ID=0x0019 | type=UINT8 | write=true | max=255 */
-            language: number;
-            /** ID=0x001a | type=BOOLEAN | write=true */
-            allCodesLockout: number;
-            /** ID=0x001b | type=BOOLEAN | write=true */
-            oneTouchLocking: number;
-            /** ID=0x001c | type=BOOLEAN | write=true */
-            privacyButtonSetting: number;
-            /** ID=0x0021 | type=UINT16 | write=true | max=65535 */
-            numberLogRecordsSupported: number;
-            /** ID=0x0030 | type=UINT8 | write=true | max=255 */
-            numberPinsSupported: number;
-            /** ID=0x0040 | type=UINT8 | write=true | max=255 */
-            numberScheduleSlotsPerUser: number;
-            /** ID=0x0050 | type=UINT8 | write=true | max=255 */
-            alarmMask: number;
-        };
-        commands: {
-            /** ID=0x10 | response=0 */
-            getLockStatus: Record<string, never>;
-            /** ID=0x12 */
-            getBatteryLevel: Record<string, never>;
-            /** ID=0x13 */
-            setRFLockoutTime: Record<string, never>;
-            /** ID=0x30 */
-            userCodeSet: {
-                /** type=CHAR_STR */
-                payload: string;
-            };
-            /** ID=0x31 */
-            userCodeGet: {
-                /** type=CHAR_STR */
-                payload: string;
-            };
-            /** ID=0x32 */
-            userCodeClear: {
-                /** type=CHAR_STR */
-                payload: string;
-            };
-            /** ID=0x33 */
-            clearAllUserCodes: Record<string, never>;
-            /** ID=0x34 */
-            setUserCodeStatus: Record<string, never>;
-            /** ID=0x35 */
-            getUserCodeStatus: Record<string, never>;
-            /** ID=0x36 */
-            getLastUserIdEntered: Record<string, never>;
-            /** ID=0x37 */
-            userAdded: Record<string, never>;
-            /** ID=0x38 */
-            userDeleted: Record<string, never>;
-            /** ID=0x40 */
-            setScheduleSlot: {
-                /** type=CHAR_STR */
-                payload: string;
-            };
-            /** ID=0x41 */
-            getScheduleSlot: {
-                /** type=CHAR_STR */
-                payload: string;
-            };
-            /** ID=0x42 */
-            setScheduleSlotStatus: {
-                /** type=CHAR_STR */
-                payload: string;
-            };
-            /** ID=0x60 | response=1 */
-            reflash: {
-                /** type=CHAR_STR */
-                payload: string;
-            };
-            /** ID=0x61 | response=2 */
-            reflashData: {
-                /** type=CHAR_STR */
-                payload: string;
-            };
-            /** ID=0x62 | response=3 */
-            reflashStatus: {
-                /** type=CHAR_STR */
-                payload: string;
-            };
-            /** ID=0x90 */
-            getReflashLock: Record<string, never>;
-            /** ID=0xa0 */
-            getHistory: Record<string, never>;
-            /** ID=0xa1 */
-            getLogin: Record<string, never>;
-            /** ID=0xa2 */
-            getUser: Record<string, never>;
-            /** ID=0xa3 */
-            getUsers: Record<string, never>;
-            /** ID=0xb0 */
-            getMandatoryAttributes: Record<string, never>;
-            /** ID=0xb1 */
-            readAttribute: Record<string, never>;
-            /** ID=0xb2 */
-            writeAttribute: Record<string, never>;
-            /** ID=0xb3 */
-            configureReporting: Record<string, never>;
-            /** ID=0xb4 */
-            getBasicClusterAttributes: Record<string, never>;
-        };
-        commandResponses: {
-            /** ID=0x00 */
-            getLockStatusRsp: {
-                /** type=UINT8 | max=255 */
-                status: number;
-            };
-            /** ID=0x01 */
-            reflashRsp: {
-                /** type=UINT8 | max=255 */
-                status: number;
-            };
-            /** ID=0x02 */
-            reflashDataRsp: {
-                /** type=UINT8 | max=255 */
-                status: number;
-            };
-            /** ID=0x03 */
-            reflashStatusRsp: {
-                /** type=UINT8 | max=255 */
-                status: number;
-            };
-        };
     };
     manuSpecificProfalux1: {
         attributes: {